Use fuller date format, updated README.
Using a better timedate resolution fixes several bugs:
Overwriting of backups, if there exists a backup with that
name (timestamp) which in effect deletes 1 backup silently.
mv error when there is only 1 current backup and it has the
name (timestamp) we're creating right now - backup gets updated
but the script exits with an error.